+++ Program.cs
@@ -23,33 +23,13 @@ static class Program
 
 
 // todo Next:
-//  + Открывать нвоый вопрос двойным щелчком по трею
-//  + Повышать вес не только слова которое не угадал, но и варианта который выбрал вместо него
-//  + Новым загруженным словам ставить автоматически рейтинг повыше-половину от максимального и сразу 5 страйков, чтобы далее уменьшалось экспоционально
-//  + Custom tray icon for paused state?
-//  + Add live preview of “next quiz in X minutes” on hover on tray
-//  + Add contributions.md file
-//  + Review ARCHITECTURE.md and validate that it matches current implementation
-//  + Improve ARCHITECTURE.md: Make sure that infrastructure layer is on the diagram
-//  + Выбирать в варианты "похожие" слова - нужно определять "расстояние" между словами
-//  ++ Возможность просить напечатать ответ (требуется нормализация списка)
-//   + IMPROVEMENTS:
-//     + Show correcly typed characters in hint (if hint is turned on)
-//     + Parametrize tests (for now only for TypingQuizPresenter)
-//     + Validate single responsibility principle for TypingQuizPresenter
-//     + Switch to Typing difficulty when reveal is set
-//     + After correct answer close window on "enter"
-//     + If answer has brackets, do not take them into account
-//     + In LetterHintTracker, if gate closed and not locked, should be a chance to reveal a bonus letter for the first option
-//     + Algorithm in SequenceAligner.FindMatches should be properly described in javadoc (or whatever they call it)
-//     + "Bonus" reveal should show random symbol, not one by one
 //  * Review namespace and directory structure
 //  * Refine words list and make sure that it is compliant
 //    * Exclude words with non-latin letters or with more than 3 words in answer (including article and words in brackets)
 //    * All multioption answers should be put correctly
 //    * Questions should not contain commas or brackets
+//  * Rewrite all testes: those which can be parametrized, should be parametrized
+//  * improve to add other languages support
 
 
 //todo Написать скрипт, чтобы смержить words и appdata, привести финальный лист в нормальный вид, убрать дубликаты если есть
-
-//todo improve to add other languages support
