+++ Services/Quiz/StringDistance.cs
@@ -1,4 +1,4 @@
-namespace VocabularyTrainer.Services;
+namespace VocabularyTrainer.Services.Quiz;
 
 public static class StringDistance
 {
