+++ ViewModels/TypingQuizViewModel.cs
@@ -1,6 +1,7 @@
 using Avalonia.Threading;
 using VocabularyTrainer.Models;
 using VocabularyTrainer.Services;
+using VocabularyTrainer.Services.Quiz;
 using Timer = System.Timers.Timer;
 
 namespace VocabularyTrainer.ViewModels;
