+++ VocabularyTrainer.Tests/Services/Quiz/HardDistractorSelectorTests.cs
@@ -1,10 +1,10 @@
 using FluentAssertions;
 using VocabularyTrainer.Models;
-using VocabularyTrainer.Services;
+using VocabularyTrainer.Services.Quiz;
 using VocabularyTrainer.Tests.Fixtures;
 using Xunit;
 
-namespace VocabularyTrainer.Tests.Services;
+namespace VocabularyTrainer.Tests.Services.Quiz;
 
 public class HardDistractorSelectorTests
 {
