+++ VocabularyTrainer.Tests/Services/Vocabulary/CsvWordRepositoryTests.cs
@@ -1,8 +1,8 @@
 using FluentAssertions;
-using VocabularyTrainer.Services;
+using VocabularyTrainer.Services.Vocabulary;
 using Xunit;
 
-namespace VocabularyTrainer.Tests.Services;
+namespace VocabularyTrainer.Tests.Services.Vocabulary;
 
 public class CsvWordRepositoryTests : IDisposable
 {
